refactor: Extract ResendCodeCountDownButton to its own file
